[rhn_channel] Add support for setting base channels
The rhn_channel module lacked the ability to change a system's base
channel. This adds the functionality and improves some of its error
handling to be more friendly.

Add an alias to parameter name, channel.

This resolves the "TypeError: cannot marshal None unless
allow_none is enabled" when the system isn't found on RHN/Satellite.